backend/requirements.txt
fastapi==0.111.0
pydantic==2.7.3
uvicorn==0.27.0
python-jose==3.3.0
passlib==1.7.4
sqlalchemy==2.0.27
fastapi-jwt-auth==0.5.0
bcrypt==4.0.1
python-multipart==0.0.7
psycopg2-binary==2.9.9
psutil